pyproject.toml
[project]
name = "code-reviewer-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-community>=0.3.8",
    "langchain>=0.3.8",
    "langgraph>=0.2.53",
    "langchain-openai>=0.2.9",
    "fastapi[standard]>=0.115.5",
    "sqlalchemy>=2.0.35",
    "pymysql>=1.1.1",
    "asyncmy>=0.2.9",
    "uvicorn>=0.32.1",
    "mysql-connector-python>=9.1.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]

[tool.uv.workspace]
members = ["moduler_app"]
